Dropped ios 5 and fixed warnings

+++ EMHintsView.m
@@ -106,7 +106,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
     CGFloat colorLocations[2] = {0.25,0.5};
     
     // draw spotlights
-    int spotlightCount = _positionArray.count;
+    NSInteger spotlightCount = _positionArray.count;
     for (int i=0; i<spotlightCount; ++i)
     {
         // center and radius of spotlight
